refactor: 절-스펙 유틸을 verse-spec.js로 분리 (bookmark.js, ADR-034 후속)
bookmark.js의 VERSE_SPEC + VERSE_SERIALIZE 블록(parseVerseSpec·selectedVersesToSpec·
mergeVerseSpecs·collapse*·serializeVerseRange)을 js/app/verse-spec.js(leaf, 의존 0)로
분리. bookmark 전용이 아니라 views(복사)·routing(parsePath 절 딥링크)도 쓰는 범용
유틸이라 별도 모듈이 맞음. bookmark.js 3,578→3,364줄.

- bookmark은 save/copy 경로용으로 명시 import. parseVerseSpec 등 외부 호출자용
  window facade는 verse-spec.js가 소유.
- bookmark.test.js의 VERSE_SPEC 슬라이스를 verse-spec.js로(extractBlock(name, source)
  일반화). index.html·sw.js 매니페스트 추가.
- bookmark.js의 export {}·appBookmark aggregate에 남아있던 옮긴 함수 제거 — 안 지우면
  `export {parseVerseSpec}`이 모듈-로컬 바인딩 없어 ESM 인스턴스화 실패(앱 전체 로드
  깨짐). tsc는 types.d.ts 전역 선언 때문에 못 잡는 사각지대 → playwright 로드 검사로 발견.

곁다리: ADR-034의 "미스노머" → "이름이 실제 내용과 안 맞음"으로 정리.

검증: tsc(main·worker) 0, 유닛 678, e2e test_navigation 12(절 딥링크 포함) + playwright
로드 검사 pageerror 0.
