+++ index.html
@@ -413,6 +413,7 @@
   <script type="module" src="/js/app/install.js"></script>
   <script type="module" src="/js/app/search.js"></script>
   <script type="module" src="/js/app/reading-context.js"></script>
+  <script type="module" src="/js/app/verse-spec.js"></script>
   <script type="module" src="/js/app/bookmark.js"></script>
   <!-- ADR-022: cite sheet (bottom sheet on mobile, right side panel on desktop) -->
   <aside id="cite-sheet" role="dialog" aria-modal="true" aria-labelledby="cite-sheet-title" hidden>
